Move the util function
